# docs/api/construction-service.openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/approval-policies/pending-platform-approval": {
      "get": {
        "description": "SYSTEM_ADMIN only, for any organization.",
        "operationId": "PolicyController_platformQueue_v1",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 512
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 25
            }
          },
          {
            "name": "workflowKey",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["project.execution", "project.completion"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["DRAFT", "PENDING_PLATFORM_APPROVAL", "ACTIVE", "REJECTED", "RETIRED"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "organizationId": {
                            "type": "string"
                          },
                          "authorOrganizationId": {
                            "type": "string"
                          },
                          "authorRole": {
                            "type": "string",
                            "enum": ["UNION_ADMIN", "SYSTEM_ADMIN"]
                          },
                          "workflowKey": {
                            "type": "string",
                            "enum": ["project.execution", "project.completion"]
                          },
                          "policyVersion": {
                            "type": "integer"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "DRAFT",
                              "PENDING_PLATFORM_APPROVAL",
                              "ACTIVE",
                              "REJECTED",
                              "RETIRED"
                            ]
                          },
                          "label": {
                            "type": "string"
                          },
                          "rationale": {
                            "type": "string"
                          },
                          "isSample": {
                            "type": "boolean"
                          },
                          "steps": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "stepOrder": {
                                  "type": "integer"
                                },
                                "approvalType": {
                                  "type": "string"
                                },
                                "authorityOrganizationId": {
                                  "type": "string"
                                },
                                "authorityRole": {
                                  "type": "string"
                                },
                                "authorityLabel": {
                                  "type": "string"
                                },
                                "minAmountMinor": {
                                  "type": "string",
                                  "nullable": true
                                },
                                "maxAmountMinor": {
                                  "type": "string",
                                  "nullable": true
                                }
                              },
                              "required": [
                                "stepOrder",
                                "approvalType",
                                "authorityOrganizationId",
                                "authorityRole",
                                "authorityLabel",
                                "minAmountMinor",
                                "maxAmountMinor"
                              ],
                              "additionalProperties": false
                            }
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "createdBy": {
                            "type": "string"
                          },
                          "submittedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "submittedBy": {
                            "type": "string",
                            "nullable": true
                          },
                          "activatedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "activatedBy": {
                            "type": "string",
                            "nullable": true
                          },
                          "rejectedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "rejectedBy": {
                            "type": "string",
                            "nullable": true
                          },
                          "rejectionReason": {
                            "type": "string",
                            "nullable": true
                          },
                          "retiredAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "retiredBy": {
                            "type": "string",
                            "nullable": true
                          },
                          "version": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "id",
                          "organizationId",
                          "authorOrganizationId",
                          "authorRole",
                          "workflowKey",
                          "policyVersion",
                          "status",
                          "label",
                          "rationale",
                          "isSample",
                          "steps",
                          "createdAt",
                          "createdBy",
                          "submittedAt",
                          "submittedBy",
                          "activatedAt",
                          "activatedBy",
                          "rejectedAt",
                          "rejectedBy",
                          "rejectionReason",
                          "retiredAt",
                          "retiredBy",
                          "version"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "nextCursor": {
                      "type": "string",
                      "nullable": true
                    },
                    "hasMore": {
                      "type": "boolean"
                    }
                  },
                  "required": ["items", "nextCursor", "hasMore"],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The request does not match the published schema. Unknown fields are refused rather than ignored, so `organizationId`, `status` or an actor field in a body is a 400. Also: an operationType outside a configured CONSTRUCTION_OPERATION_TYPES list, and an operating area PostGIS considers invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No credentials, or a token that is expired, unverifiable or issued for another audience.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted: a role the configuration does not grant (INSUFFICIENT_ROLE), the oversight role, a service-to-service token, a SYSTEM_ADMIN that has not selected an organization with X-Organization-Id, on a decision, a caller who can see the project but is not the authority the approval names; on an approval policy, an author who is not a union or platform administrator, a union writing for an organization not beneath it, a caller other than the author organization submitting or retiring it, a non-SYSTEM_ADMIN approving or rejecting it, or the person who wrote or submitted it approving it (four eyes; a union-written policy always); on requesting approval or completing, a union-written policy in force whose union no longer governs the organization (re-confirmed at use).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "summary": "The platform administrator's queue: every organization's policies awaiting approval",
        "tags": ["approval-policies"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/approval-policies/{id}": {
      "get": {
        "description": "Visible to its author organization, to the governed organization's project readers and to a platform administrator; anyone else gets 404.",
        "operationId": "PolicyController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "authorOrganizationId": {
                      "type": "string"
                    },
                    "authorRole": {
                      "type": "string",
                      "enum": ["UNION_ADMIN", "SYSTEM_ADMIN"]
                    },
                    "workflowKey": {
                      "type": "string",
                      "enum": ["project.execution", "project.completion"]
                    },
                    "policyVersion": {
                      "type": "integer"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "DRAFT",
                        "PENDING_PLATFORM_APPROVAL",
                        "ACTIVE",
                        "REJECTED",
                        "RETIRED"
                      ]
                    },
                    "label": {
                      "type": "string"
                    },
                    "rationale": {
                      "type": "string"
                    },
                    "isSample": {
                      "type": "boolean"
                    },
                    "steps": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "stepOrder": {
                            "type": "integer"
                          },
                          "approvalType": {
                            "type": "string"
                          },
                          "authorityOrganizationId": {
                            "type": "string"
                          },
                          "authorityRole": {
                            "type": "string"
                          },
                          "authorityLabel": {
                            "type": "string"
                          },
                          "minAmountMinor": {
                            "type": "string",
                            "nullable": true
                          },
                          "maxAmountMinor": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "stepOrder",
                          "approvalType",
                          "authorityOrganizationId",
                          "authorityRole",
                          "authorityLabel",
                          "minAmountMinor",
                          "maxAmountMinor"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "createdBy": {
                      "type": "string"
                    },
                    "submittedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "submittedBy": {
                      "type": "string",
                      "nullable": true
                    },
                    "activatedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "activatedBy": {
                      "type": "string",
                      "nullable": true
                    },
                    "rejectedAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "rejectedBy": {
                      "type": "string",
                      "nullable": true
                    },
                    "rejectionReason": {
                      "type": "string",
                      "nullable": true
                    },
                    "retiredAt": {
                      "type": "string",
                      "nullable": true
                    },
                    "retiredBy": {
                      "type": "string",
                      "nullable": true
                    },
                    "version": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "authorOrganizationId",
                    "authorRole",
                    "workflowKey",
                    "policyVersion",
                    "status",
                    "label",
                    "rationale",
                    "isSample",
                    "steps",
                    "createdAt",
                    "createdBy",
                    "submittedAt",
                    "submittedBy",
                    "activatedAt",
                    "activatedBy",
                    "rejectedAt",
                    "rejectedBy",
                    "rejectionReason",
                    "retiredAt",
                    "retiredBy",
                    "version"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "The request does not match the published schema. Unknown fields are refused rather than ignored, so `organizationId`, `status` or an actor field in a body is a 400. Also: an operationType outside a configured CONSTRUCTION_OPERATION_TYPES list, and an operating area PostGIS considers invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "No credentials, or a token that is expired, unverifiable or issued for another audience.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted: a role the configuration does not grant (INSUFFICIENT_ROLE), the oversight role, a service-to-service token, a SYSTEM_ADMIN that has not selected an organization with X-Organization-Id, on a decision, a caller who can see the project but is not the authority the approval names; on an approval policy, an author who is not a union or platform administrator, a union writing for an organization not beneath it, a caller other than the author organization submitting or retiring it, a non-SYSTEM_ADMIN approving or rejecting it, or the person who wrote or submitted it approving it (four eyes; a union-written policy always); on requesting approval or completing, a union-written policy in force whose union no longer governs the organization (re-confirmed at use).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Not found — also returned for a project, need, progress report, policy or approval that belongs to another organization (and, for an approval, whose authority the caller is not), so its existence is never disclosed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "summary": "Read one approval policy with its steps",
        "tags": ["approval-policies"],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Rasta — Construction Service",
    "description": "Civil-works projects (CON-001): drafting with needs, configurable approvals, execution, progress reports and completion. Approval authorities come from approval_policy rows the union writes and the platform administrator approves (Q-70 (7)) — an (organization, role) per step, in order, by estimate range — never from code, and the platform never creates an authority or approves by default: no policy, no applicable step, silence or a timeout is a refusal, and only the authority a step names may decide it (ADR-023, ADR-063). Every change is a compare-and-set on `version` and publishes its event on rasta.construction.v1 in the same transaction. Project fields, roles, preconditions, start and completion rules, and progress rules are provisional answers to docs/24 Q-68 to Q-73, each configurable. Every read and write is confined to the organization the request acts for, except that an authority of another organization sees and decides the steps addressed to it.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {}
          }
        },
        "required": ["code", "message"],
        "additionalProperties": false
      }
    }
  }
}